# internal/docs/services/auth/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and session management",
        "title": "Auth Service API",
        "contact": {
            "name": "Hossein Asadian",
            "email": "Hosseinasadian442@email.com"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Authenticate user and return tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1.4.2",
                        "description": "App version",
                        "name": "X-App-Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "iPhone 15 Pro",
                        "description": "Device name",
                        "name": "X-Device-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Installation UUID",
                        "name": "X-Installation-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            "ios",
                            "android",
                            "web"
                        ],
                        "type": "string",
                        "description": "Platform",
                        "name": "X-Platform",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/refresh-token": {
            "post": {
                "description": "Refresh access token using refresh token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh Token",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1.4.2",
                        "description": "App version",
                        "name": "X-App-Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "iPhone 15 Pro",
                        "description": "Device name",
                        "name": "X-Device-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Installation UUID",
                        "name": "X-Installation-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            "ios",
                            "android",
                            "web"
                        ],
                        "type": "string",
                        "description": "Platform",
                        "name": "X-Platform",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Refresh token credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "example": "1.4.2",
                        "description": "App version",
                        "name": "X-App-Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "iPhone 15 Pro",
                        "description": "Device name",
                        "name": "X-Device-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Installation UUID",
                        "name": "X-Installation-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            "ios",
                            "android",
                            "web"
                        ],
                        "type": "string",
                        "description": "Platform",
                        "name": "X-Platform",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Register credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "push_token": {
                    "type": "string"
                }
            }
        },
        "auth.LoginResponse": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/auth.User"
                },
                "user_access_token": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshTokenRequest": {
            "type": "object",
            "properties": {
                "push_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "push_token": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterResponse": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "user_access_token": {
                    "type": "string"
                }
            }
        },
...
        "auth.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}